Resolves #26.
While placeholder techniques such as LQIP or SQIP do adequately fill the space, I'm no longer of the mind that they're an effective technique from a psychological or performance standpoint. So I have removed references to these techniques altogether. Use an inlined data URI to an extremely small placeholder, or a use a solid color backdrop in CSS.